requirements.txt
aiohttp==3.7.4.post0
humanfriendly==10.0
jishaku==2.3.2
py-cord==2.0.0b3
python-dotenv==0.19.2
async-timeout==3.0.1
requests==2.27.1